Add missing class attributes and reorder properties
Added missing class attributes needed for other subsystems like eg. file_path in Solution. Reordered some of the properties of classes and edited their not-nullity – eg. Administrator has to have username and password, but doesn't have to have first or last name, so the required fields should be before the additional ones.